# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 风电检测资料管理系统 API
  version: 0.1.0
  description: 公司内网资料管理系统后端 API
paths:
...
  /api/v1/documents/{id}/:
    get:
      operationId: v1_documents_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: 标识此 文档 的 唯一整数值。
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
    put:
      operationId: v1_documents_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: 标识此 文档 的 唯一整数值。
        required: true
      tags:
      - v1
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DocumentUpdate'
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
    patch:
      operationId: v1_documents_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: 标识此 文档 的 唯一整数值。
        required: true
      tags:
      - v1
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedDocumentUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedDocumentUpdate'
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedDocumentUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
...
  /api/v1/documents/{id}/permanent-delete/:
    post:
      operationId: v1_documents_permanent_delete_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: 标识此 文档 的 唯一整数值。
        required: true
      tags:
      - v1
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentMutation'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DocumentMutation'
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentMutation'
        required: true
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/documents/{id}/versions/:
    post:
      operationId: v1_documents_versions_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: 标识此 文档 的 唯一整数值。
        required: true
      tags:
      - v1
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentVersionUpload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DocumentVersionUpload'
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentVersionUpload'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentVersion'
          description: ''
  /api/v1/documents/batch-download/:
    post:
      operationId: v1_documents_batch_download_create
      tags:
      - v1
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentBatchDownload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DocumentBatchDownload'
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentBatchDownload'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: string
                format: binary
          description: ''
...
components:
  schemas:
    AccessLevelEnum:
      enum:
      - internal
      - restricted
      type: string
      description: |-
        * `internal` - 内部
        * `restricted` - 受限
...
    Document:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        project:
          type: integer
          readOnly: true
          nullable: true
          title: 所属项目
        project_name:
          type: string
          readOnly: true
        folder:
          type: integer
          readOnly: true
          title: 所属文件夹
        folder_name:
          type: string
          readOnly: true
        title:
          type: string
          readOnly: true
          title: 文档标题
        description:
          type: string
          readOnly: true
          title: 描述
        access_level:
          allOf:
          - $ref: '#/components/schemas/AccessLevelEnum'
          readOnly: true
          title: 访问级别
        current_version:
          allOf:
          - $ref: '#/components/schemas/DocumentVersion'
          readOnly: true
        can_download:
          type: boolean
          readOnly: true
        can_update:
          type: boolean
          readOnly: true
        can_delete:
          type: boolean
          readOnly: true
        can_restore:
          type: boolean
          readOnly: true
        can_create_version:
          type: boolean
          readOnly: true
        lock_version:
          type: integer
          readOnly: true
          title: 乐观锁版本
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: 删除时间
        deleted_by:
          type: integer
          readOnly: true
          nullable: true
          title: 删除人
        deleted_by_name:
          type: string
          readOnly: true
        created_by:
          type: integer
          readOnly: true
          nullable: true
          title: 创建人
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: 更新时间
      required:
      - access_level
      - can_create_version
      - can_delete
      - can_download
      - can_restore
      - can_update
      - created_at
      - created_by
      - created_by_name
      - current_version
      - deleted_at
      - deleted_by
      - deleted_by_name
      - description
      - folder
      - folder_name
      - id
      - lock_version
      - project
      - project_name
      - title
      - updated_at
    DocumentBatchDownload:
      type: object
      properties:
        document_ids:
          type: array
          items:
            type: integer
            minimum: 1
          maxItems: 20
          minItems: 1
      required:
      - document_ids
...
    DocumentMutation:
      type: object
      properties:
        expected_updated_at:
          type: string
          format: date-time
      required:
      - expected_updated_at
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        access_level:
          $ref: '#/components/schemas/AccessLevelEnum'
        expected_updated_at:
          type: string
          format: date-time
      required:
      - expected_updated_at
...
    DocumentVersion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        document:
          type: integer
          readOnly: true
          title: 文档
        version_number:
          type: integer
          readOnly: true
          title: 版本号
        original_filename:
          type: string
          readOnly: true
          title: 原始文件名
        content_type:
          type: string
          readOnly: true
          title: 内容类型
        file_size:
          type: integer
          readOnly: true
          title: 文件大小
        sha256:
          type: string
          readOnly: true
          title: SHA-256
        uploaded_by:
          type: integer
          readOnly: true
          nullable: true
          title: 上传人
        uploaded_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
      required:
      - content_type
      - created_at
      - document
      - file_size
      - id
      - original_filename
      - sha256
      - uploaded_by
      - uploaded_by_name
      - version_number
    DocumentVersionUpload:
      type: object
      properties:
        file:
          type: string
          format: uri
      required:
      - file
...
    PatchedDocumentUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        access_level:
          $ref: '#/components/schemas/AccessLevelEnum'
        expected_updated_at:
          type: string
          format: date-time
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid